pyproject.toml
[tool.poetry]
name = "llm-notes"
version = "0.1.0"
description = ""
authors = ["Leo Browning <leobrowning92@gmail.com>"]
readme = "README.md"
packages = [{include = "llm_notes"}]


[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.4"
pydantic = "^1.10.7"
asyncio = "^3.4.3"
pyyaml = "^6.0"
langchain = "^0.0.186"
openai = "^0.27.7"
wrapt = "^1.15.0"
unstructured = "^0.6.11"
chromadb = "^0.3.25"
tiktoken = "^0.4.0"
networkx = "^3.1"
guidance = "^0.0.64"
faiss-cpu = "^1.7.4"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^3.6.3"
black = "^23.3.0"
flake8 = "^6.0.0"
matplotlib = "^3.7.1"
plotly = "^5.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
